// eatfit247-admin/libs/shared-ui/src/lib/empty-state/empty-state.enum.ts
export enum EmptyStateType {
  MEMBER_POCKET_GUIDE = 'member_pocket_guide',
  MEMBER_HEALTH_ISSUES = 'member_health_issues',
  MEMBER_ASSESSMENT = 'member_assessment',
  MEMBER_CALL_LOGS = 'member_call_logs',
  MEMBER_PAYMENT_HISTORY = 'member_payment_history',
  MEMBER_DIET_HISTORY = 'member_diet_history',
  MEMBER_BODY_STATS = 'member_body_stats',
  MEMBER_DASHBOARD = 'member_dashboard',
  GENERAL = 'general',
}